Move rpc_context and test_response classes into their own files.
The file nano/rpc_test/common.hpp was very confusing, a lot of stuff
was dumped into one file and it was hard to read.
This helps break things into more manageable pieces.
