+++ nano/rpc_test/CMakeLists.txt
@@ -1,4 +1,14 @@
-add_executable(rpc_test common.hpp common.cpp entry.cpp receivable.cpp rpc.cpp)
+add_executable(
+  rpc_test
+  common.hpp
+  rpc_context.hpp
+  test_response.hpp
+  rpc_context.cpp
+  test_response.cpp
+  common.cpp
+  entry.cpp
+  receivable.cpp
+  rpc.cpp)
 
 target_link_libraries(rpc_test test_common)
